+++ tests/integration/querier_sqlite/testdata/030_named_parameter_at/runner.go
@@ -39,7 +39,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	user, err := queries.GetUserByID(ctx, int32(1))
+	user, err := queries.GetUserByID(ctx, int64(1))
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
